Genericized BoundsPlane and BoundsMesh to work with more than just AMVs
